+++ app/controllers/events_controller.rb
@@ -1,7 +1,7 @@
 class EventsController < ApplicationController
   def index
     all_events = Event.where(end_date: Time.zone.today..).or(Event.where(start_date: Time.zone.today.., end_date: nil))
-    @events = all_events
+    @events = all_events.includes(event_attendees: [:user])
     @events = all_events.where(country_code: params[:country]) if params[:country].present?
     @years = @events.order(:start_date)
                     .group_by { |e| e.start_date.year }