+++ db/migrate/20251021121357_add_event_attendences.rb
@@ -1,0 +1,12 @@
+class AddEventAttendences < ActiveRecord::Migration[8.0]
+  def change
+    create_table :event_attendances do |t|
+      t.references :user, null: false, foreign_key: true
+      t.references :event, null: false, foreign_key: true
+
+      t.timestamps
+    end
+
+    add_index :event_attendances, [:user_id, :event_id], unique: true
+  end
+end