+++ db/migrate/20251021132624_rename_attendances_to_attendees.rb
@@ -1,0 +1,5 @@
+class RenameAttendancesToAttendees < ActiveRecord::Migration[8.0]
+  def change
+    rename_table :event_attendances, :event_attendees
+  end
+end