+++ test/controllers/events_controller_test.rb
@@ -1,0 +1,8 @@
+require 'test_helper'
+
+class EventsControllerTest < ActionDispatch::IntegrationTest
+  test 'index' do
+    get events_url
+    assert_response :success
+  end
+end